backend/requirements.txt
django
djangorestframework
environs
drf-yasg